requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.11.0
dataclasses==0.8
dbus-python==1.2.16
distro-info==0.23+ubuntu1.1
idna==2.8
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-contrib-python==4.1.0.25
pandas==1.1.5
Pillow==8.2.0
PyGObject==3.36.0
pyparsing==3.1.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.2.0
six==1.14.0
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.64.1
typing_extensions==4.1.1
unattended-upgrades==0.1
urllib3==1.25.8
zipp==3.6.0
